requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
werkzeug
python-dotenv
marshmallow
flask-marshmallow
marshmallow-sqlalchemy
flask-jwt-extended
flask_bcrypt
Flask-Migrate